Don't depend on test load order
